update requirements.txt, move file list functions to local_git.py
