+++ backend/cli/__init__.py
@@ -1,0 +1,1 @@
+from .local_git import *
